+++ app/Http/Controllers/Dashboard/Pickups/Controller.php
@@ -15,6 +15,7 @@ use App\Logic\Dashboard\CRUD\Services\Pickups as PickupsService;
 use App\Logic\Dashboard\CRUD\Repositories\Pickups as PickupsRepository;
 
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 
 final class Controller extends Controllers
 {
@@ -49,6 +50,12 @@ final class Controller extends Controllers
      */
     public function index(PickupsRequest $request): JsonResponse
     {
+        if(! Gate::check('Pickups')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'filters' => $this->service->getAllFilters($request),
 
@@ -69,6 +76,12 @@ final class Controller extends Controllers
      */
     public function store(PickupsRequest $request): JsonResponse
     {
+        if(! Gate::check('Pickups')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'message' => 'The pickup was succesfully created.',
 
@@ -83,6 +96,12 @@ final class Controller extends Controllers
      */
     public function show(Pickup $pickup): JsonResponse
     {
+        if(! Gate::check('Pickups')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'pickup' => $this->service->showPickup($pickup),
         ]);
@@ -97,6 +116,12 @@ final class Controller extends Controllers
      */
     public function update(PickupsRequest $request, Pickup $pickup): JsonResponse
     {
+        if(! Gate::check('Pickups')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         return Json::sendJsonWith200([
             'message' => 'The pickup was successfully updated.',
 
@@ -111,6 +136,12 @@ final class Controller extends Controllers
      */
     public function destroy($id): JsonResponse
     {
+        if(! Gate::check('Pickups')){
+            return Json::sendJsonWith403([
+                'message' =>  'Permission denied.'
+            ]);
+        }
+
         $id = $this->service->deletePickup($id);
 
         if(!$id)
